% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.single.example}
\alias{df.single.example}
\title{infectionAgeHIV Example Data}
\format{
## `df.single.example`
A data frame with 4 rows and 4 columns.
Contains values for which there is only measurement per patient
\describe{
  \item{patient_id}{A unique identifier for the patient}
  \item{last_neg_test_date}{The patient's last negative HIV test date, if available}
  \item{first_pos_test_date}{The patient's first positive HIV test date}
  \item{ART_start_date}{The date the patient started antiretroviral therapy}
}
}
\usage{
df.single.example
}
\description{
Example data for use in the vignette for infectionAgeHIV
}
\keyword{datasets}
