% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.infection.ages.R
\name{find.infection.ages}
\alias{find.infection.ages}
\title{Find infection age distributions}
\usage{
find.infection.ages(
  df.single,
  df.multiple,
  n.adapt = 10000,
  n.burn = 10000,
  n.iter = 1e+05
)
}
\arguments{
\item{df.single}{A data frame of last negative test date, first positive test date, and ART start date for each patient}

\item{df.multiple}{A data frame of biomarker values for each patient}

\item{n.adapt}{Number of iterations to adapt the model}

\item{n.burn}{Number of iterations to run before sampling starts}

\item{n.iter}{Number of iterations of sampling after burn-in is completed}
}
\value{
A data frame with the inputs, quality controlled inputs (removing values too long after ART start),
and inferred distributions
}
\description{
Function to determine which biomarker values are usable in the multiple biomarker model and find distributions
for the amount of time between infection and diagnosis
}
