% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_biomarker_model.R
\name{mbm.predict}
\alias{mbm.predict}
\title{Predict infection times with MBM}
\usage{
mbm.predict(
  BED = rep(NA, 1),
  LAg = rep(NA, 1),
  CD4 = rep(NA, 1),
  pol = rep(NA, 1),
  pol2 = rep(NA, 1),
  prev.neg.time = NULL,
  t.BED.delay = rep(0, length(BED)),
  t.LAg.delay = rep(0, length(LAg)),
  t.CD4.delay = rep(0, length(CD4)),
  t.pol.delay = rep(0, length(pol)),
  t.pol2.delay = rep(0, length(pol2)),
  mub = NULL,
  Sigmab = NULL,
  sigmae = NULL,
  n.adapt = 1000,
  n.burn = 1000,
  n.iter = 1000,
  prior.type = NULL,
  inf.mean = 2,
  inf.sd = 1.5,
  max.seroconvert.delay = 2/12,
  u1.pdf = NULL,
  output.raw = FALSE,
  seed = sample(1e+09, 1),
  ...
)
}
\arguments{
\item{BED}{A vector of BED values, one for each individual. Can be omitted.}

\item{LAg}{A vector of LAg values, one for each individual. Can be omitted.}

\item{CD4}{A vector of CD4 values, one for each individual. These are the actual CD4+ T-cell concentrations, not the square-root transformed ones used in previous versions. 
Can be omitted.}

\item{pol}{A vector for the proportion of polymorphic sites in the HIV polymerase gene, one for each individual.
Can be omitted.}

\item{pol2}{A vector of values for the diversity in the HIV polymerase gene as obtained from NGS, one for each individual. Can be omitted.}

\item{prev.neg.time}{The amount of time in between an individual's positive HIV test and a previous negative test. 
Use Inf or NA if no previous negative test is available.}

\item{t.BED.delay}{A vector or matrix for the difference in time (in years (days/365.25)) between the dates of the positive test and when BED was sampled.}

\item{t.LAg.delay}{A vector or matrix for the difference in time (in years (days/365.25)) between the dates of the positive test and when LAg was sampled.}

\item{t.CD4.delay}{A vector or matrix for the difference in time (in years (days/365.25)) between the dates of the positive test and when CD4 was sampled.}

\item{t.pol.delay}{A vector or matrix for the difference in time (in years (days/365.25)) between the dates of the positive test and when pol was sampled.}

\item{t.pol2.delay}{A vector or matrix for the difference in time (in years (days/365.25)) between the dates of the positive test and when pol2 was sampled.}

\item{mub}{A vector of length 12 for the means of the random effects parameters for the Multiple Biomarker Model.}

\item{Sigmab}{A 12x12 matrix for the covariance matrix of the random effects parameters for the Multiple Biomarker Model}

\item{sigmae}{A vector for the variances (not standard deviations) of the error paramemeters in the Multiple Biomarker Model}

\item{n.adapt}{The number of iterations to adapt the rjags MCMC model}

\item{n.burn}{The number of iterations of burn-in for the rjags MCMC model}

\item{n.iter}{The number of iterations of samples for the rjags MCMC model.}

\item{prior.type}{The type of prior used for the time between infection and diagnosis in the multiple biomarker model.
1 indicates a gamma distribution with mean `inf.mean` years and standard deviation `inf.sd` years.
2 indicates a continuous uniform distribution with minimum 0 and maximum 12 years.
3 indicates a distribution for an individual that is HIV positive but has not developed AIDS symptoms, 
assuming that AIDS symptoms develop after a length of time according to a gamma distribution with shape 3.349 and rate 0.327
4 indicates a user-supplied distribution from the parameter user.prior.pdf}

\item{inf.mean}{The mean of the gamma distribution used for the prior distribution when there is no previous negative test. Default value is 2 years.}

\item{inf.sd}{The mean of the gamma distribution used for the prior distribution when there is no previous negative test. Default value is 1.5 years.}

\item{max.seroconvert.delay}{The maximum reasonable about of time that someone can be infected before an HIV test would be positive.
It must be a list with x and y components corresponding to the time before diagnosis and probability density.}

\item{output.raw}{If TRUE, the output will include all MCMC samples for the infection ages.}

\item{seed}{The RNG seed to use.}
}
\value{
The probability density of the infection times for each patient. 
  "pdf" is a continuous function and "pdf.num" is a density class object as produced by the density function.
  If output.raw is TRUE, "raw" is the matrix of MCMC samples of the infection ages for all individuals.
}
\description{
Function to use the multiple biomarker model for infection time prediction
This version uses the actual CD4 values, not the square root transformed values of previous versions
}
