% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_HIV_data.R
\name{prepare.HIV.data}
\alias{prepare.HIV.data}
\title{Prepare data for/use multiple biomarker model}
\usage{
prepare.HIV.data(
  patient_ID,
  last_neg_test_date = NULL,
  first_pos_test_date = NULL,
  ART_start_date = NULL,
  BED_dates = NULL,
  BED = NULL,
  LAg_dates = NULL,
  LAg = NULL,
  CD4_dates = NULL,
  CD4 = NULL,
  seq_dates,
  seqs,
  seq_names = NULL,
  pol_override = NULL,
  seq_length_override = NULL,
  pol2_dates = NULL,
  pol2 = NULL,
  VL_dates = NULL,
  VL = NULL,
  ART_cutoff_delay = 3,
  cluster_ID = NULL,
  gender = NULL,
  age_at_sampling = NULL,
  birth_location = NULL,
  suspected_infection_location = NULL,
  risk_group = NULL,
  aids_diagnosis_date = NULL,
  death_date = NULL,
  date_format = "\%Y-\%m-\%d",
  find_infection_age_distributions = FALSE,
  prior.type = 1,
  user.prior.pdf = list(x = c(0, 10), y = c(1/10, 1/10)),
  n.adapt = 10000,
  n.burn = 1e+05,
  n.iter = 1e+06,
  seed = sample(2^31 - 1, 1),
  ...
)
}
\arguments{
\item{patient_ID}{An identifier for each patient}

\item{last_neg_test_date}{The last negative test date, if available}

\item{first_pos_test_date}{The first positive test date. It is overriden by the earliest biomarker sample date if that is earlier
or if the first positive test date is unavailable.}

\item{ART_start_date}{The start date of antiretroviral therapy}

\item{BED_dates}{A list of vectors for the sample dates of the BED values}

\item{BED}{A list of vectors for values for the BED test}

\item{LAg_dates}{A list of vectors for the sample dates of the LAg values}

\item{LAg}{A list of vectors for values for the LAg test}

\item{CD4_dates}{A list of vectors for the sample dates of the CD4 counts}

\item{CD4}{A list of vectors for the CD4+ T-cell counts}

\item{seq_dates}{A list of vectors for the sample dates of the HIV sequences}

\item{seq_names}{The names of the sequences}

\item{pol_override}{A list of vectors for values for the polymorphism count if they have already been calculated externally. 
These values will replace the pol values calculated by this function.}

\item{seq_length_override}{A list of vectors for the sequence lengths 
(to be used along with pol_override if they have already been calculated externally)}

\item{pol2_dates}{A list of vectors for the sample dates of the pol2 values}

\item{VL_dates}{A list of vectors for the sample dates of the viral load}

\item{VL}{A list of vectors for values for the viral load}

\item{ART_cutoff_delay}{The amount of time (in days) after the start of antiretroviral therapy 
after which biomarker values should be disregarded}

\item{cluster_ID}{An identifier for the transmission cluster that the patient belongs to}

\item{gender}{A vector for the genders of each patient}

\item{age_at_sampling}{A list of vectors for the age of each patient at the time of each sequence}

\item{birth_location}{A vector for the birth location of each patient}

\item{suspected_infection_location}{A vector for the suspected infection location of each patient}

\item{risk_group}{A vector for the transmission risk group of each patient}

\item{aids_diagnosis_date}{A vector for the AIDS diagnosis date of each patient if they have been diagnosed with AIDS}

\item{death_date}{The date of death of each patient if the patient has died}

\item{date_format}{A character string specifying the format of the dates (to be passed to as.Date())}

\item{find_infection_age_distributions}{Whether or not to run the multiple biomarker model (MBM) on the data}

\item{prior.type}{The type of prior used for the time between infection and diagnosis in the multiple biomarker model.
1 indicates a gamma distribution with mean 2 years and standard deviation 1.5 years.
2 indicates a continuous uniform distribution with minimum 0 and maximum 12 years.
3 indicates a distribution for an individual that is HIV positive but has not developed AIDS symptoms, 
assuming that AIDS symptoms develop after a length of time according to a gamma distribution with shape 3.349 and rate 0.327
4 indicates a user-supplied distribution from the parameter user.prior.pdf}

\item{user.prior.pdf}{A pdf to use as the prior distribution for the time between infection and diagnosis in the multiple biomarker model.
It must be a list with x and y components corresponding to the time before diagnosis and probability density.}

\item{n.adapt}{Number of adaptation iterations if the MBM is run}

\item{n.burn}{Number of burn-in iterations if the MBM is run}

\item{n.iter}{Number of sampling iterations if the MBM is run}

\item{seed}{The RNG seed}

\item{...}{Additional lists of patient data to put into the dataframe}

\item{seq}{A list of HIV sequences (or a list of lists of sequences) for each patient in character or DNAbin format}

\item{pol}{A list of vectors for values for the pol2 value}
}
\value{
A dataframe containing the input data, quality controlled versions of the data, possibly infection age distributions
}
\description{
Function to put biomarker and other date into a dataframe, calculate the polypmorphism count,
and (possibly) run the multiple biomarker model on it
}
