% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_biomarker_model.R
\name{sim.biomarkers}
\alias{sim.biomarkers}
\title{Simulate biomarker values}
\usage{
sim.biomarkers(
  t.inf,
  mub = NULL,
  Sigmab = NULL,
  sigmae = NULL,
  which.biomarkers = matrix(1, nrow = length(t.inf), ncol = 5),
  seed = sample(1e+09, 1)
)
}
\arguments{
\item{t.inf}{A vector or list of the true infection ages for each individual. 
If given as a list, multiple infection ages can used per individual, and the output will be returned as a list.}

\item{mub}{A vector of length 12 for the means of the random effects parameters for the Multiple Biomarker Model.}

\item{Sigmab}{A 12x12 matrix for the covariance matrix of the random effects parameters for the Multiple Biomarker Model}

\item{sigmae}{A vector for the variances (not standard deviations) of the error paramemeters in the Multiple Biomarker Model}

\item{which.biomarkers}{A vector of length 5 or matrix of size (number of individuals)x5 for the biomarkers to simulate for each individual.
If 1, the corresponding biomarker will be simulated. If 0, the corresponding biomarker will not be simulated.
The order of the biomarkers is: BED, LAg, CD4, pol, and pol2.
If a vector is given, the same set of biomarkers will be used for each individual (although the actual biomarker values are gemerated independently)}

\item{seed}{The RNG seed to use.}
}
\value{
A Matrix of simulated biomarkers from the supplied model parameters. The order of the biomarkers is: BED, LAg, CD4, pol, and pol2.
}
\description{
Function to simulate biomarkers
This version uses the actual CD4 values, not the square root transformed values of previous versions
}
